requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2
Faker==15.3.4
importlib-resources==5.10.0
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.31
PyJWT==2.5.0
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.4
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.9.0
